backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.8
pydantic-settings>=2.4
sqlmodel>=0.0.22
psycopg[binary]>=3.2
python-multipart>=0.0.9
pandas>=2.2
openpyxl>=3.1
rapidfuzz>=3.9
python-dateutil>=2.9
openai>=1.0
httpx>=0.27
passlib[bcrypt]>=1.7

